requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosmtplib==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1